Rename to `k-search-bar`
