--- panel/lab/components/search/index.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-	'docs' => 'k-search',
-];
