+++ panel/lab/components/search/search-bar/index.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+	'docs' => 'k-search-bar',
+];
