+++ panel/lab/components/search/search-bar/index.vue
@@ -1,25 +1,25 @@
 <template>
 	<k-lab-examples>
 		<k-lab-example label="Default">
-			<k-search />
+			<k-search-bar />
 		</k-lab-example>
 		<k-lab-example label="Types">
-			<k-search :types="types" />
+			<k-search-bar :types="types" />
 		</k-lab-example>
 		<k-lab-example label="Default type">
-			<k-search :types="types" default-type="files" />
+			<k-search-bar :types="types" default-type="files" />
 		</k-lab-example>
 		<k-lab-example label="No results">
-			<k-search :results="[]" />
+			<k-search-bar :results="[]" />
 		</k-lab-example>
 		<k-lab-example label="Results">
-			<k-search :results="results" />
+			<k-search-bar :results="results" />
 		</k-lab-example>
 		<k-lab-example label="More results">
-			<k-search :results="results" :pagination="{ total: 20 }" />
+			<k-search-bar :results="results" :pagination="{ total: 20 }" />
 		</k-lab-example>
 		<k-lab-example label="Is Loading">
-			<k-search :is-loading="true" :results="results" :types="types" />
+			<k-search-bar :is-loading="true" :results="results" :types="types" />
 		</k-lab-example>
 	</k-lab-examples>
 </template>
