+++ panel/src/components/Navigation/SearchBar.vue
